requirements.txt
# FastAPI stack
fastapi==0.115.0
uvicorn[standard]==0.30.0  # [standard] geeft extra features

# Database (async voor FastAPI)
asyncpg==0.29.0  # Vervangt psycopg2-binary

# Configuratie
python-dotenv>=1.0.1

pydantic==2.8.0
pydantic-settings==2.1.0